# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_research_pdf_endpoint_research_pdf_post": {
        "properties": {
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_research_pdf_endpoint_research_pdf_post",
        "type": "object"
      },
...
      "DeleteResponse": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "deleted",
          "id"
        ],
        "title": "DeleteResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResearchInput": {
        "properties": {
          "depth": {
            "default": "medium",
            "enum": [
              "shallow",
              "medium",
              "deep"
            ],
            "title": "Depth",
            "type": "string"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "model_overrides": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Overrides"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "topic"
        ],
        "title": "ResearchInput",
        "type": "object"
      },
      "ResearchReplyInput": {
        "properties": {
          "model_overrides": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Overrides"
          },
          "reply": {
            "title": "Reply",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "reply"
        ],
        "title": "ResearchReplyInput",
        "type": "object"
      },
      "ResearchSessionListResponse": {
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/ResearchSessionSummary"
            },
            "title": "Sessions",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "sessions",
          "total",
          "limit",
          "offset"
        ],
        "title": "ResearchSessionListResponse",
        "type": "object"
      },
      "ResearchSessionRenameInput": {
        "properties": {
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "topic"
        ],
        "title": "ResearchSessionRenameInput",
        "type": "object"
      },
      "ResearchSessionSummary": {
        "properties": {
          "coverage_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Pct"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "depth": {
            "title": "Depth",
            "type": "string"
          },
          "domain": {
            "title": "Domain",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "iterations_completed": {
            "title": "Iterations Completed",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          },
          "total_entries_ingested": {
            "title": "Total Entries Ingested",
            "type": "integer"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "topic",
          "status",
          "depth",
          "domain",
          "iterations_completed",
          "total_entries_ingested",
          "created_at",
          "updated_at"
        ],
        "title": "ResearchSessionSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Self-hosted RAG-powered workflow orchestrator",
    "title": "Scaffold Engine",
    "version": "1.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/research": {
      "post": {
        "description": "Autonomous research: decompose topic → search → extract → ingest → iterate.\n\nWrapped in ``_sse_with_disconnect_watch`` so that client disconnect\npropagates a ``CancelledError`` into the research generator within ~1s,\nallowing the lifecycle wrapper to finalize the session as ``cancelled``.",
        "operationId": "research_endpoint_research_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Research Endpoint",
        "tags": [
          "Research"
        ]
      }
    },
    "/research/pdf": {
      "get": {
        "description": "Drag-and-drop HTML upload page for PDF ingestion.",
        "operationId": "research_pdf_upload_page_research_pdf_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Research Pdf Upload Page",
        "tags": [
          "Research"
        ]
      },
      "post": {
        "description": "PDF ingestion: upload PDF → extract → ingest → stream SSE.",
        "operationId": "research_pdf_endpoint_research_pdf_post",
        "parameters": [
          {
            "in": "query",
            "name": "extractor",
            "required": false,
            "schema": {
              "default": "auto",
              "pattern": "^(auto|pypdf|plumber)$",
              "title": "Extractor",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "domain",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Domain"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_research_pdf_endpoint_research_pdf_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Research Pdf Endpoint",
        "tags": [
          "Research"
        ]
      }
    },
    "/research/reply": {
      "post": {
        "description": "Resume a paused research session with the user's clarification reply.",
        "operationId": "research_reply_endpoint_research_reply_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchReplyInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Research Reply Endpoint",
        "tags": [
          "Research"
        ]
      }
    },
    "/research/sessions": {
      "get": {
        "description": "Paginated research session list with optional status + topic search.",
        "operationId": "list_research_sessions_research_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchSessionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Research Sessions",
        "tags": [
          "Management"
        ]
      }
    },
    "/research/sessions/{session_id}": {
      "delete": {
        "description": "Hard-delete a research session. Note: KB entries already in Milvus are NOT\nremoved; this only drops the session metadata + state snapshot.",
        "operationId": "delete_research_session_research_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Delete Research Session",
        "tags": [
          "Management"
        ]
      },
      "patch": {
        "description": "Rename a research session (set topic).",
        "operationId": "rename_research_session_research_sessions__session_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchSessionRenameInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchSessionSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Rename Research Session",
        "tags": [
          "Management"
        ]
      }
    },
    "/research/verify/{session_id}": {
      "get": {
        "description": "Session-scoped provenance audit (§17.114 + §17.121).\n\nLists every Milvus entry produced by the given research session and\nreports its current state — present, superseded, or missing. Used to\nsurface drift between what was ingested and what's currently in the\nindex, without re-fetching upstream content. See\n``app/modules/research_verify.py`` for the returned-shape contract.\n\n``?recheck=true`` (§17.121) additionally HEAD-requests each entry's\n``source_url`` and reports ``upstream_state`` (reachable / missing /\nforbidden / error / skipped) per entry plus rollup totals. Bounded\nconcurrency (5). SSRF re-checked on every URL.\n\nPre-§17.114 sessions have no provenance rows linked by session_id\nand return an empty ``entries`` list — that's expected, not an error.",
        "operationId": "research_verify_endpoint_research_verify__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "description": "If true, HEAD-request each entry's source_url to surface upstream reachability state.",
            "in": "query",
            "name": "recheck",
            "required": false,
            "schema": {
              "default": false,
              "description": "If true, HEAD-request each entry's source_url to surface upstream reachability state.",
              "title": "Recheck",
              "type": "boolean"
            }
          },
          {
            "description": "If true (§17.126), GET each URL and SHA256-compare against the stored raw_upstream_hash. Implies recheck=true.",
            "in": "query",
            "name": "compare_hash",
            "required": false,
            "schema": {
              "default": false,
              "description": "If true (§17.126), GET each URL and SHA256-compare against the stored raw_upstream_hash. Implies recheck=true.",
              "title": "Compare Hash",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Research Verify Endpoint",
        "tags": [
          "Research"
        ]
      }
    },
...
  }
}